+++ src/chart/options/config.h
@@ -16,10 +16,7 @@ public:
 	static std::string paramsJson();
 	[[nodiscard]] std::string getParam(const std::string &path) const;
 	void setParam(const std::string &path, const std::string &value);
-	explicit Config(Options &options, Data::DataTable &table) :
-	    options(options),
-	    table(table)
-	{}
+	explicit Config(Options &options, Data::DataTable &table);
 
 private:
 	std::reference_wrapper<Options> options;